Setting up test dependency in pom.xml and implementing springboot/BidTests.java
